adds contract ids for account history, moves and cleans up test helpers
