--- src/stub.test.ts
@@ -1,5 +1,0 @@
-describe("Stub test", () => {
-  it("passes as placeholder", () => {
-    expect(true).toBeTruthy;
-  });
-});
